Working (#2)
* deleted handlebars folders

* add userRoutes

* Created Home.js

Created the Home.js file for front end

* Created Search.js

Created Search.js file for gif searching through front end

* Created NotFound.js

Created the page for bad search request for front end

* Created App.js

Created the App.js file to include routing and page components. Still need to add login, logout, and signup to it once those files are completed

* updated dependencies

* minor changes in pages

* updated userRoutes lonin

* added heroku link to README

* Created SignUp.js

Created Sign Up page structure for application

* Created Login.js

Created the Login page structure for the application

* Created Logout.js

Created the Logout page structure for the application

* Updated App.js

Updated the App.js file to include the newly added Login, Logout and Sign Up pages files for routing

* Created mutations.js

Created the mutations file to be able to update data with

* Created queries.js

Created the queries file to be able to define search queries

* Updated Logout.js

Updated the Logout.js file to be able to utilize the mutations.js file

* Updated Login.js

Updated the Login.js file to be able to utilize the mutations.js file

* Updated SignUp.js

Updated the SignUp.js file to be able to utilize the mutations.js file

* Updated Search.js

Updated the Search.js file to be able to utilize the queries.js file

* Updated Home.js

Updated the Home.js file to have button links to the Login, Logout and Sign Up pages and to link search queries through the home page to the separate search page.

* updated userRoutes login part

* updated userRoutes logout part

* Created package.json

Created the client side package.json for the application

* Created .gitignore

Created the .gitignore for the client side

* Created auth.js

Created the auth.js for JWT authentication to use in the pages files

* Updated Home.js

Updated the Home.js to import the auth.js and utilize authentication

* Updated Login.js

Updated the Login.js to import the auth.js and utilize the authentication

* Updated Logout.js

Updated the Logout.js to import the auth.js and utilize the authentication

* Updated Search.js

Updated the Search.js file to require people to be logged in to use the search function by importing and utilizing authentication from the auth.js file.

* Created index.js

Created the index.js file for the client side folder

* Created index.css

Borrowed the index.css from prior classes for this application

* Created index.html

Created the entry point for the application

* Created manifest.json

Created the manifest.json file

* Created robots.txt

Created the robots.txt file

* Updated userRoutes.js

Updated the URL paths to matchup with front end App.js file

* Deleted config file to move to a server folder

* Created connection.js

Still need to recode to make this work for MERN application but moved it to appropriate folder

* Deleted folder to move to server folder

Needs to be in correct file/folder structure

* Moved to correct folder structure

Still need to recode to make this work for MERN application but moved it to appropriate folder

* Moved to correct folder

* Deleted sql database

We're using MongooseDB

* Deleted to move to correct folder structure

* Needs to be in correct file/folder structure

Still need to recode to make this work for MERN application but moved it to appropriate folder

* Needs to be in correct file/folder structure

Still need to recode to make this work for MERN application but moved it to appropriate folder

* Needs to be in correct file/folder structure

* Created App.css

Moved to correct folder

* Needs to be in correct file/folder structure

* Needed to be in correct file/folder structure

* Deleted handlebars

These files were substituted into src folder as pages files

* Needs to be in correct file/folder structure

* Needed to be in correct file/folder structure

Still need to recode to make this work for MERN application but moved it to appropriate folder

* Deleted .gitignore

Already have this in client folder

* Deleted index.html

Already in client folder

* Needs to be in correct file/folder structure

* Needs to be in correct file/folder structure

Still need to recode to make this work for MERN application but moved it to appropriate folder

* Updated Home.js

changed div to header

* Updated Gif.js

Updated file to work for MERN application

* Updated User.js

Updated file to work for MERN application

* Created index.js

Created the index file for the models folder

* formated syntax, updated README

* branchy merge, updated dependencies

* Updated Search.js

Took out redundant h2

* Deleted models folder

Doesn't follow file structure. This folder is already in the server folder.

* Deleted public folder and css file

Doesn't follow file structure. This folder is also already in the client folder.

* Deleted views folder

Doesn't follow file structure. This folder is already in the client folder in the pages folder.

* Deleted node modules folder

We don't need the node modules in our repo

* Updated index.js

Updated the file to work for MERN application

* Created package.json

Created the package.json for the server folder

* Created manifest.json

Copied and pasted from the manifest.json in the public folder

* Created index.html

Copy and pasted from the public folder

* Created robots.txt

Copy and pasted from the public folder

* Updated connection.js

Updated the file to work with MERN application

* updated signIn.js

* Updated .env

Updated the file to work with MERN application

---------

Co-authored-by: Guangshi Jie <guangshi18@gmail.com>
Co-authored-by: Kyle Kryst <119367684+KyleKryst@users.noreply.github.com>